bbgdp plotter

+++ code/cpp/src/util/Config.h
@@ -33,8 +33,8 @@ struct PhysicsConfig {
 
 struct SolverConfig {
   double tolerance;
-  int max_bbpgd_iterations;
-  int max_recursive_iterations;
+  size_t max_bbpgd_iterations;
+  size_t max_recursive_iterations;
   double linked_cell_size;
   double growth_factor;
   double particle_preallocation_factor;
